+++ api_docs/changelog.md
@@ -20,6 +20,13 @@ format used by the Zulip server that they are interacting with.
 
 ## Changes in Zulip 10.0
 
+**Feature level 312**
+
+* [`GET /invites/{invite_id}`](/api/get-invite-details): Endpoint response includes details of
+  the email invitation by invite id.
+* [`GET /invites/multiuse/{invite_id}`](/api/get-invite-link-details): Endpoint response includes
+  the details of the reusable invitation link by invite id.
+
 **Feature level 311**
 
 * [`POST /user_groups/{user_group_id}/members`](/api/update-user-group-members):